pyproject.toml
[tool.poetry]
name = "llms-co-writer"
version = "1.0.0"
description = "The LLMs Co-Writer API is a powerful RESTful service that leverages the prowess of large language models (LLMs), such as OpenAI's GPT, to assist in textual completion, co-writing, and other creative tasks. "
license = "Proprietary"
authors = ["Mikheil Tabidze <m.tabidze@gmail.com>"]
readme = "README.md"
repository = "https://https://github.com/mtabidze/llms-co-writer"
packages = [{include = "llms_co_writer"}]


[tool.poetry.dependencies]
python = "^3.11"
pip = "^23.3.1"
fastapi = "^0.104.1"
pydantic-settings = "^2.1.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
openai = "^1.3.6"
redis = "^5.0.1"
boto3 = "^1.33.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-random-order = "^1.1.0"
pytest-asyncio = "^0.20.3"
coverage = "^7.3.2"
mock = "^5.1.0"
bandit = "^1.7.5"
pip-audit = "^2.6.1"
pip-licenses = "^4.3.3"
flake8 = "^6.1.0"
flake8-bugbear = "^23.11.28"
flake8-copyright = "^0.2.4"
pylama = "^8.4.1"
ruff = "^0.0.290"
black = "^23.11.0"
isort = "^5.12.0"


[tool.isort]
profile = "black"


[tool.pylama.linter.mccabe]
max-complexity = 16


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
